third_party/blink/web_tests/display-lock/lock-before-append/measure-forced-svg-text.html

<!doctype HTML>

<!--
Runs an acquire, which constructs and measures a subtree.
This should do a forced layout. Commit also measures it to get
post-final-layout values.
-->

<style>
#container {
  contain: style layout;
  width: 100px;
  height: 100px;
  background: lightgreen;
}
</style>

<div id="parent"></div>

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>

<script>
async_test((t) => {
  function measureForced() {
    t.step(() => {
      assert_not_equals(document.getElementById("text").getComputedTextLength(), 0, "forced");
    });
  }

  function measureInCommit() {
    t.step(() => {
      assert_not_equals(document.getElementById("text").getComputedTextLength(), 0, "in commit");
    });
  }

  function construct(container) {
    container.innerHTML = `
      <svg viewBox="0 0 100 100" xmlns="http://www.w3.org/2000/svg">
        <style>
          .t { font: 10px sans-serif; }
        </style>
        <text id="text" x="10" y="10" class="t">This is text</text>
      </svg>
    `;
  }

  async function runTest() {
    let container = document.createElement("div");
    container.id = "container";
    await container.displayLock.acquire({ timeout: Infinity });

    document.getElementById("parent").appendChild(container);
    construct(container);
    measureForced();

    container.displayLock.commit().then(() => {
      measureInCommit();
      t.done();
    });
  }

  window.onload = function() {
    requestAnimationFrame(() => requestAnimationFrame(runTest));
  };
}, "Measure Forced SVG Text");
</script>
